I PostgresqlGroupDAO - реализована логика I PostgresqlContactDAO - реализована логика I PostgresqlUserDAO - реализована логика I PostgresqlAnalyticalInfoDAO - реализована логика

+++ src/main/java/ru/bellintegrator/app/dao/impl/sql/AnalyticalInfoDAO.java
@@ -10,9 +10,9 @@ public interface AnalyticalInfoDAO extends GenericDAO<AnalyticalInfo>{
 
     int getUserCount() throws DAOException;
 
-    Map<Integer, Integer> getEachUserContactCount() throws DAOException;
+    Map<Integer, Long> getEachUserContactCount() throws DAOException;
 
-    Map<Integer, Integer> getEachUserGroupCount() throws DAOException;
+    Map<Integer, Long> getEachUserGroupCount() throws DAOException;
 
     double getAvgUserCountInGroup() throws DAOException;
